refactor / cleanup a bit
